Source code push
Probs the only thing i'll be doing in this repo. Only other thing would be to add some(lots of) code that saves the texture offline or draws it to the screen (but what flavor of api to use on what platform will become a not fun mess so yeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee).